% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_data_validation.R
\name{add_date_year_month}
\alias{add_date_year_month}
\title{Create date from year and month}
\usage{
add_date_year_month(df)
}
\arguments{
\item{df}{data.frame with columns "Monat" and "Jahr" (both numeric)}
}
\description{
Create date from year and month
}
