% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_observations.R
\name{extract.observations}
\alias{extract.observations}
\title{Extract for measured traffic intensity and velocity}
\usage{
\method{extract}{observations}(x, ...)
}
\arguments{
\item{x}{single_source}

\item{...}{other parameters}
}
\description{
load the traffic data from the web
}
