% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_data_validation.R
\name{number_to_weekday}
\alias{number_to_weekday}
\title{Transform number to weekday abbreviation}
\usage{
number_to_weekday(x)
}
\arguments{
\item{x}{a number}
}
\description{
For example, 1 is "Mo", 0 is "So", but 7 is "So" too
}
