% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_single_sources.R
\name{data_sources}
\alias{data_sources}
\alias{single_source}
\title{Data sources to query from}
\usage{
data_sources()

single_source(
  name,
  get_data = read.xlsx,
  send_data = fairqDbtools::send_data,
  load_mode = "replace",
  load_table = paste0("traffic_det_", name),
  ...
)
}
\arguments{
\item{name}{(character) a unique name used as class and to identify target
table on database.}

\item{get_data}{(function)}

\item{send_data}{(function)}

\item{load_mode}{(character) the mode in which we load data.}

\item{load_table}{(character) the table in which we load data.}

\item{...}{arguments used within methods and available for queries}
}
\description{
List data sources. This list defines which tables to extract.
}
