% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_cross_sections.R
\name{transform.cross_sections}
\alias{transform.cross_sections}
\title{transform cross sections data (Messquerschnitte)}
\usage{
\method{transform}{cross_sections}(x, ...)
}
\arguments{
\item{x}{single_source}

\item{...}{other parameters}
}
\description{
Bus lanes and left turn lanes are excluded because they
introduce a bias. They usually have low numbers of vehicles although they may
be in heavy traffic areas. That confuses the traffic model.

transforms the cross sections metadata
}
