% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_observations.R
\name{transform.observations}
\alias{transform.observations}
\title{Transform for measured traffic intensity and velocity}
\usage{
\method{transform}{observations}(x, ...)
}
\arguments{
\item{x}{single_source}

\item{...}{other parameters}
}
\description{
rename columns for traffic observations
}
