% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_data_validation.R
\name{vehicle_types_to_row}
\alias{vehicle_types_to_row}
\title{Make data longer}
\usage{
vehicle_types_to_row(df, names_prefix, cols)
}
\arguments{
\item{df}{data.frame with traffic detector data}

\item{names_prefix}{Name for the new value column; this prefix plus "_" is removed from the
names of the columns filling the value column}

\item{cols}{columns to fill the new long value column}
}
\description{
Make data longer
}
